Combine pkg/sys/mounter and pkg/sys/mounter/k8smounter
Combines the two packages in a single one as there is no much
value on keeping them separate and makes the project structure
innecessarily convoluted.

Originally there were separated to keep interface definition
and implementation separate to make possible importing
the interface without having to necessarily import the
implementation.
